third_party/blink/web_tests/wpt_internal/trust-tokens/trust-token-api-rate-limiting.https.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Tests trust token issuance and redemption</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/get-host-info.sub.js"></script>
<script>
    'use strict';

    function issue_token() {
        return fetch('/wpt_internal/trust-tokens/resources/trust_token_issuance.py', {
            privateToken: { type: 'private-state-token', version: 1, operation: 'token-request' }
        });
    }

    function redeem_token() {
        return fetch('/wpt_internal/trust-tokens/resources/trust_token_redemption.py', {
            privateToken: { type: 'private-state-token', version: 1, operation: 'token-redemption', refreshPolicy: "refresh" }
        });
    }

    promise_test(async function () {
        // Issue a batch of tokens
        const batch_size = 10;
        const requests = [];
        for (let i = 0; i < batch_size; i++) {
            requests.push(issue_token());
        }
        await Promise.all(requests).then(responses =>
            responses.forEach(response => assert_equals(response.status, 200)));

        // Redeem three tokens - the last attempt should fail due to rate limiting
        var response = await redeem_token();
        assert_equals(response.status, 200);
        response = await redeem_token();
        assert_equals(response.status, 200);
        return promise_rejects_dom(this, 'OperationError', redeem_token());

    }, 'Token redemption is rate limited.');

</script>
